requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13,<4.0
mysqlclient>=2.0,<3.0  # For MySQL database support
xhtml2pdf>=0.2,<1.0  # For PDF generation
python-decouple>=3.5,<4.0  # For environment variable management
django-cors-headers>=3.10,<4.0  # For handling CORS headers
django-filter>=22.1,<23.0  # For filtering querysets
